Update 4
Added repository query.  Added Repository component that returns a paragraph showing the repo name url and a list showing the last 5 issues.  Added functionality to onSubmit and onChange methods within the main app component to set the state path to the input text and call the onFetchFromGithub method when the submit button is pressed.  The onFetchFromGithub method parses the input url into organization and repository variables then makes a post request to Github and returns the data result to the respective state variables within the app component for use in rendering.