Update 12
Added buttons for each comment that adds heart emoticon to a comment on click,  Added has reaction function within comment component.  The function loops through all the comment reactions and returns true if the user has reacted to a comment.  The buttons display a message to add a heart or remove heart depending on the result of the has reaction comment.